v3/requirements.txt
asyncio==3.4.3
sqlalchemy==1.4.45
python-dotenv==0.21.0
asyncpg==0.27.0
psycopg2==2.9.5
uvicorn==0.20.0
fastapi==0.88.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.5
python-jose==3.3.0
pydantic==1.10.2